// src/parser/descriptors/generated/userCmdDeltaSchema.ts
export const userCmdDeltaSchema = {
	CSGOUserCmdPB: {
		fields: {
			1: { wire: 2, child: 'CBaseUserCmdPB' },
			2: { wire: 2, child: 'CSGOInputHistoryEntryPB', repeated: true },
			6: { wire: 0, def: [255, 255, 255, 255, 255, 255, 255, 255, 255, 1] },
			7: { wire: 0, def: [255, 255, 255, 255, 255, 255, 255, 255, 255, 1] },
			9: { wire: 0, def: [0] },
			11: { wire: 0, def: [0] },
			12: { wire: 0, def: [0] },
			13: { wire: 0, def: [0] }
		}
	},
	CBaseUserCmdPB: {
		fields: {
			1: { wire: 0, def: [0] },
			2: { wire: 0, def: [0] },
			3: { wire: 2, child: 'CInButtonStatePB' },
			4: { wire: 2, child: 'CMsgQAngle' },
			5: { wire: 5, def: [0, 0, 0, 0] },
			6: { wire: 5, def: [0, 0, 0, 0] },
			7: { wire: 5, def: [0, 0, 0, 0] },
			8: { wire: 0, def: [0] },
			9: { wire: 0, def: [0] },
			10: { wire: 0, def: [0] },
			11: { wire: 0, def: [0] },
			12: { wire: 0, def: [0] },
			14: { wire: 0, def: [255, 255, 255, 7] },
			17: { wire: 0, def: [0] },
			18: { wire: 2, child: 'CSubtickMoveStep', repeated: true },
			19: { wire: 2 },
			20: { wire: 0, def: [0] },
			21: { wire: 0, def: [0] },
			22: { wire: 2, child: 'CBaseUserCmdExecutionNotes' }
		}
	},
	CInButtonStatePB: {
		fields: {
			1: { wire: 0, def: [0] },
			2: { wire: 0, def: [0] },
			3: { wire: 0, def: [0] }
		}
	},
	CMsgQAngle: {
		fields: {
			1: { wire: 5, def: [0, 0, 0, 0] },
			2: { wire: 5, def: [0, 0, 0, 0] },
			3: { wire: 5, def: [0, 0, 0, 0] }
		}
	},
	CSubtickMoveStep: {
		fields: {
			1: { wire: 0, def: [0] },
			2: { wire: 0, def: [0] },
			3: { wire: 5, def: [0, 0, 0, 0] },
			4: { wire: 5, def: [0, 0, 0, 0] },
			5: { wire: 5, def: [0, 0, 0, 0] },
			8: { wire: 5, def: [0, 0, 0, 0] },
			9: { wire: 5, def: [0, 0, 0, 0] }
		}
	},
	CBaseUserCmdExecutionNotes: {
		fields: {
			1: { wire: 2 }
		}
	},
	CSGOInputHistoryEntryPB: {
		fields: {
			2: { wire: 2, child: 'CMsgQAngle' },
			4: { wire: 0, def: [0] },
			5: { wire: 5, def: [0, 0, 0, 0] },
			6: { wire: 0, def: [0] },
			7: { wire: 5, def: [0, 0, 0, 0] },
			12: { wire: 2, child: 'CSGOInterpolationInfoPB_CL' },
			13: { wire: 2, child: 'CSGOInterpolationInfoPB' },
			14: { wire: 2, child: 'CSGOInterpolationInfoPB' },
			15: { wire: 2, child: 'CSGOInterpolationInfoPB' },
			64: { wire: 0, def: [0] },
			65: { wire: 0, def: [255, 255, 255, 255, 255, 255, 255, 255, 255, 1] },
			66: { wire: 2, child: 'CMsgVector' },
			67: { wire: 2, child: 'CMsgVector' },
			68: { wire: 2, child: 'CMsgVector' },
			69: { wire: 2, child: 'CMsgQAngle' }
		}
	},
	CSGOInterpolationInfoPB_CL: {
		fields: {
			3: { wire: 5, def: [0, 0, 0, 0] }
		}
	},
	CSGOInterpolationInfoPB: {
		fields: {
			1: { wire: 0, def: [255, 255, 255, 255, 255, 255, 255, 255, 255, 1] },
			2: { wire: 0, def: [255, 255, 255, 255, 255, 255, 255, 255, 255, 1] },
			3: { wire: 5, def: [0, 0, 0, 0] }
		}
	},
	CMsgVector: {
		fields: {
			1: { wire: 5, def: [0, 0, 0, 0] },
			2: { wire: 5, def: [0, 0, 0, 0] },
			3: { wire: 5, def: [0, 0, 0, 0] },
			4: { wire: 5, def: [0, 0, 0, 0] }
		}
	}
}